+++ database/migrations/2022_08_16_070844_kategori_final.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('kode_utility_kategori');
             $table->string('kode_kriteria');
             $table->string('nim');
-            $table->decimal('nilai_akhir');
+            $table->float('nilai_akhir', 9, 5);
             $table->timestamps();
         });
     }
